+++ src/features/badge/index.ts
@@ -1,3 +1,10 @@
+//ui
 export { BadgeHeader } from './ui/BadgeHeader';
 export { BadgeStats } from './ui/BadgeStats';
 export { BadgeTabs } from './ui/BadgeTabs';
+
+//type
+export type { BadgeResponse, BadgeType, BaseBadge, EarnedBadge } from './types/badge';
+
+// api 훅
+export { useBadges } from './api/useBadges';
