+++ src/features/badge/types/badge.ts
@@ -1,26 +1,3 @@
-// import type { LucideIcon } from 'lucide-react';
-
-// export interface BadgeType {
-//   id: number;
-//   name: string;
-//   description: string;
-//   icon: LucideIcon;
-//   color: string;
-//   status?: 'earned' | 'in-progress' | 'locked';
-//   earnedDate?: string;
-//   category: string;
-//   rarity: string;
-//   progress?: number;
-//   target?: number;
-//   requirement?: s
-
-// export interface BadgeTabsProps {
-//   earnedBadges: BadgeType[];
-//   progressBadges: BadgeType[];
-//   lockedBadges: BadgeType[];
-//   allBadges: BadgeType[];
-// }
-
 // 공통 배지 타입
 export interface BaseBadge {
   name: string;
